requirements.txt
loguru==0.7.3
pytest==9.0.2
PyYAML==6.0.3